Assignment function
PJAC does not support value reassignment in the strict sense.
You can return a value from functio call into a variable (that's the
only way to "reassign" a value to a variable).

With "auto" as a return type came an interesting construct:

    function assign(auto x) -> auto { return x; }
    function f() {
        var int a = 42;
        var int b = 64;
        a = assign(b);
    }

After the "a = assign(b);" line, "a" has the same value as "b".
You gotta love hacks...

However, this creates a hole in the type system and
needs to be fixed.
And by "fixed" I mean *functions with "auto" parameters should be
considered templates and have special rules for type checking".
A special case is a function with "auto" parameter and "auto" return
type as it is a kind of "uber-template".

The "assign(auto)->auto" is the only use I currently found for it.